refactor: put the msales and orders page in folders
